% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{scribe}
\alias{scribe}
\title{Create a Specification for Simulation}
\usage{
scribe(..., simulator)
}
\arguments{
\item{...}{Simulator specifications.}

\item{simulator}{The name of the simulator, which will be used as the class name for
the returned charm object.}
}
\description{
Create a Specification for Simulation
}
