% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{conjure}
\alias{conjure}
\title{Simulate Data Using Provided Specs}
\usage{
conjure(incantation, ...)
}
\arguments{
\item{incantation}{An `incantation` object.}

\item{...}{Optional additional arguments.}
}
\description{
Simulate Data Using Provided Specs
}
