% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{scribe}
\alias{scribe}
\title{Create a Specification for Simulation}
\usage{
scribe(specs, simulator, ...)
}
\arguments{
\item{specs}{A named list.}

\item{simulator}{The name of the simulator, which will be used as the class name for}

\item{...}{Optional additional arguments
the returned incantation object.}
}
\description{
Create a Specification for Simulation
}
